requirements.txt
asgiref==3.2.7
Django==3.0.5
django-cors-headers==3.2.1
django-environ==0.4.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
